multi-provider/requirements.txt
openai>=1.0.0,<3.0.0
anthropic>=0.40.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
